[Feat] 카메라 촬영 후 동작 변경

+++ app/src/main/java/com/codehanzoom/greenwalk/view/HomeScreen.kt
@@ -163,7 +163,7 @@ fun areaHeader() {
 fun areaMyInfo(name: String?="나희수",
                ploggingCount: Int?=-1,
                totalPoint: Int?=-1,
-               totalWalkingDistance: Int?=-1,
+               totalWalkingDistance: Double?=-1.0,
                grade: String?="Bronze") {
     Card(
         elevation = CardDefaults.cardElevation(
